fix: moved searchable dropdown to global components
