+++ packages/ui/src/rebrand/dropdown/SearchableDropdown.tsx
@@ -1,6 +1,7 @@
 import {Dispatch, MutableRefObject, ReactNode, SetStateAction, useEffect, useRef} from "react";
-import { DropdownArrow } from "../../../../../../packages/ui/src/rebrand/icons/DropArrowIcon";
-import { WarningIcon } from "../../../../../../packages/ui/src/rebrand/icons/WarningIcon";
+import { DropdownArrow } from "../icons/DropArrowIcon";
+import { WarningIcon } from "../icons/WarningIcon";
+
 
 
 
@@ -32,7 +33,7 @@ interface DropdownItemProps {
     children: ReactNode;
 }
 
-export const DropdownItem = ({onClick, extraClasses, key, dropdownOptionsCount, children}: DropdownItemProps) => {
+export const SearchableDropdownItem = ({onClick, extraClasses, key, dropdownOptionsCount, children}: DropdownItemProps) => {
     return <p onClick={onClick} className={`flex items-center px-[15px] hover:bg-abaddonBlack bg-black cursor-pointer duration-300 ease-in-out text-lg min-h-[48px] font-medium ${dropdownOptionsCount > DROPDOWN_ITEMS_WITHOUT_SCROLL && "mr-[2px]"} ${extraClasses}`} key={key}>{children}</p>
 }
