+++ packages/ui/src/rebrand/dropdown/index.ts
@@ -1,3 +1,4 @@
-import {Dropdown, DropdownItem} from "./Dropdown";
+import { Dropdown, DropdownItem } from "./Dropdown";
+import { SearchableDropdown, SearchableDropdownItem } from "./SearchableDropdown";
 
-export {Dropdown, DropdownItem}
+export {Dropdown, DropdownItem, SearchableDropdown, SearchableDropdownItem}
